FINISHED, testing strncmp

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: smendez- <smendez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 12:42:02 by smendez-          #+#    #+#             */
-/*   Updated: 2024/11/10 14:07:22 by smendez-         ###   ########.fr       */
+/*   Updated: 2024/11/15 11:01:42 by smendez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-int	ft_strlen(const char *c)
+size_t	ft_strlen(const char *c)
 {
 	size_t	i;
